Refactor : 숫자 예외를 체크하는 메소드를 NumberCheck로 변경
